FWA-50 - Config settings are getting reset on refresh
